fix: delegate subcounter and enforce type preference
The logic for keeping track of subscriber count for observable objects
has been removed from the `subscriptions` table and moved to a seperate
`subcounter` table that mirrors it in structure and simply keeps a
tally. This fixed a bug where publishing data or modifying it would do
nothing.

The publish method has been given preference for functions as arguments,
same as that for the observer pattern.